+++ internal/transcribe/types.go
@@ -1,5 +1,7 @@
 package speech
 
+import "context"
+
 type audioInfo struct {
 	sampleRate  int
 	numChannels int
@@ -12,6 +14,10 @@ type audioProcessor interface {
 
 type audioProcessorFactory interface {
 	createProcessor(fileExtension string) (audioProcessor, error)
+}
+
+type transcriber interface {
+	process(ctx context.Context, filename string, opts TranscribeOptions) (string, error)
 }
 
 type defaultAudioProcessorFactory struct{}
